requirements.txt
pillow
numpy
matplotlib
scipy
tensorflow
pandas
rich
opencv-python
cvzone